requirements.txt
wheel==0.26.0
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
binaryornot==0.4.0
certifi==2016.2.28
chardet==2.3.0
complexity==0.9.1
dominate==2.2.0
Flask==0.10.1
Flask-Bootstrap==3.3.5.7
flask-nav==0.4
Flask-Script==2.0.5
Flask-WTF==0.12
itsdangerous==0.24
Jinja2==2.8
livereload==2.3.2
Markdown==2.3.1
MarkupSafe==0.23
PyYAML==3.11
singledispatch==3.4.0.3
six==1.10.0
tornado==4.3
Twisted==16.1.1
visitor==0.1.2
Werkzeug==0.11.9
WTForms==2.1
requests==2.10.0
zope.interface==4.1.3
xlutils==2.0.0
Flask-SQLAlchemy==2.1
celery==3.1.24
msgpack-python==0.4.8
babel==2.3.4
flower==0.9.1
futures==3.0.5
tornado==4.3
PyMySQL==0.8.1
#MySQL-python==1.2.5